CRUD item

+++ SpringBoot/src/main/java/itu/mbds/transversal/repository/ItemCategoryRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemCategoryRepo extends JpaRepository<ItemCategory, Integer> {
+public interface ItemCategoryRepo extends JpaRepository<ItemCategory, Long> {
 
     Page<ItemCategory> findByTitleContainsIgnoreCase(String keyword, Pageable pageable);
 }
